+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Karafka Framework Changelog
 
-## 2.4.10 (Unreleased)
+## 2.4.10 (2024-09-02)
 - **[Feature]** Provide Kafka based Scheduled Messages to be able to send messages in the future via a proxy topic.
 - [Enhancement] Introduce a `#assigned` hook for consumers to be able to trigger actions when consumer is built and assigned but before first consume/ticking, etc.
 - [Enhancement] Provide `Karafka::Messages::Message#tombstone?` to be able to quickly check if a message is a tombstone message.
 - [Enhancement] Provide more flexible API for Recurring Tasks topics reconfiguration.
 - [Enhancement] Remove no longer needed Rails connection releaser.
+- [Enhancement] Update AppSignal client to support newer versions (tombruijn and hieuk09).
 - [Fix] Fix a case where there would be a way to define multiple subscription groups for same topic with different consumer.
 
 ## 2.4.9 (2024-08-23)